+++ src/ArduinoJson/Object/ObjectRef.hpp
@@ -57,6 +57,7 @@ class ObjectRefBase {
 };
 
 class ObjectConstRef : public ObjectRefBase<const CollectionData>,
+                       public VariantOperators<ObjectConstRef>,
                        public Visitable {
   friend class ObjectRef;
   typedef ObjectRefBase<const CollectionData> base_type;
@@ -147,6 +148,7 @@ class ObjectConstRef : public ObjectRefBase<const CollectionData>,
 
 class ObjectRef : public ObjectRefBase<CollectionData>,
                   public ObjectShortcuts<ObjectRef>,
+                  public VariantOperators<ObjectRef>,
                   public Visitable {
   typedef ObjectRefBase<CollectionData> base_type;